+++ core/src/test/scala/org/specs2/specification/ExecutorSpec.scala
@@ -136,10 +136,10 @@ class ExecutorSpec extends script.Specification with Groups with ResultMatchers 
     val timeFactor = env.arguments.execute.timeFactor
 
     val messages = new ListBuffer[String]
-    def verySlow      = { Thread.sleep(600 * timeFactor); messages.append("very slow"); success }
+    def verySlow      = { Thread.sleep(600 * timeFactor.toLong); messages.append("very slow"); success }
 
     val fragments = Seq(example("very slow", verySlow))
-    val env1 = env.setTimeout(100.millis * timeFactor)
+    val env1 = env.setTimeout(100.millis * timeFactor.toLong)
 
     execute(fragments, env1) must contain(beSkipped[Result]("timeout after "+100*timeFactor+" milliseconds"))
   }
@@ -156,12 +156,12 @@ class ExecutorSpec extends script.Specification with Groups with ResultMatchers 
     // this cannot be made lazy vals otherwise this will block on 'slow'
     def ok(name: String)              = {                    messages.append(name); success }
     def fast(timeFactor: Int)         = {                    messages.append("fast"); success }
-    def medium(timeFactor: Int)       = { Thread.sleep(10 * timeFactor);  messages.append("medium"); success }
+    def medium(timeFactor: Int)       = { Thread.sleep(10 * timeFactor.toLong);  messages.append("medium"); success }
     def ex(s: String)                 = { messages.append(s); success }
-    def mediumFail(timeFactor: Int)   = { Thread.sleep(10 * timeFactor);  messages.append("medium"); failure }
-    def mediumSkipped(timeFactor: Int)= { Thread.sleep(10 * timeFactor);  messages.append("medium"); skipped }
-    def slow(timeFactor: Int)         = { Thread.sleep(200 * timeFactor); messages.append("slow");   success }
-    def verySlow(timeFactor: Int)     = { Thread.sleep(600 * timeFactor); messages.append("very slow"); success }
+    def mediumFail(timeFactor: Int)   = { Thread.sleep(10 * timeFactor.toLong);  messages.append("medium"); failure }
+    def mediumSkipped(timeFactor: Int)= { Thread.sleep(10 * timeFactor.toLong);  messages.append("medium"); skipped }
+    def slow(timeFactor: Int)         = { Thread.sleep(200 * timeFactor.toLong); messages.append("slow");   success }
+    def verySlow(timeFactor: Int)     = { Thread.sleep(600 * timeFactor.toLong); messages.append("very slow"); success }
     def step1                         = { messages.append("step");   success }
     def fatalStep                     = { messages.append("fatal");  if (true) throw new java.lang.Error("fatal error!"); success }
   }
